+++ lib/hoodoo/services/middleware/middleware.rb
@@ -214,7 +214,7 @@ module Hoodoo; module Services
     #   of a search string, would find records on-or-after the date.
     #
     # Values are either a validation Proc or +nil+ for no validation. The
-    # Proc takes the search query value as its sole input paraeter and must
+    # Proc takes the search query value as its sole input parameter and must
     # evaluate to the input value either unmodified or in some canonicalised
     # form if it is valid, else to +nil+ if the input value is invalid. The
     # canonicalisation is typically used to coerce a URI query string based
@@ -1160,8 +1160,11 @@ module Hoodoo; module Services
       # Compile the remaining log payload and send it.
 
       unless secure
-        body = interaction.rack_request.body.read( MAXIMUM_LOGGED_PAYLOAD_SIZE )
-               interaction.rack_request.body.rewind()
+        body = nil
+        if !interaction.rack_request.body.nil?
+          body = interaction.rack_request.body.read( MAXIMUM_LOGGED_PAYLOAD_SIZE )
+                 interaction.rack_request.body.rewind()
+        end
 
         data[ :payload ][ :body ] = body
       end
@@ -1342,7 +1345,7 @@ module Hoodoo; module Services
     end
 
     # This is part of the formalised structured logging interface upon which
-    # external entites might depend. Change with care.
+    # external entities might depend. Change with care.
     #
     # For a given interaction, log the response *after the fact* of calling
     # a resource implementation, using the target interface's resource name
@@ -1373,7 +1376,7 @@ module Hoodoo; module Services
       # For other kinds of data, check the secure actions to see if the body
       # should be included.
       #
-      # TODO: This uses deprecated acccessors into the "context.request.list"
+      # TODO: This uses deprecated accessors into the "context.request.list"
       #       object, but it keeps the code simple. It'd be nice to just have
       #       e.g. "data[ :list ] = context.request.list.to_h()" but the
       #       change in log output format might break dependent clients.
@@ -1416,7 +1419,7 @@ module Hoodoo; module Services
     end
 
     # This is part of the formalised structured logging interface upon which
-    # external entites might depend. Change with care.
+    # external entities might depend. Change with care.
     #
     # For a given service interface, an implementation of which is receiving
     # a given action under the given request context, log the response *after
@@ -1495,7 +1498,7 @@ module Hoodoo; module Services
     # Log a debug message. Pass optional extra arguments which will be used as
     # strings that get appended to the log message.
     #
-    # THIS IS INSCURE. Sensitive data might be logged. DO NOT USE IN DEPLOYED
+    # THIS IS INSECURE. Sensitive data might be logged. DO NOT USE IN DEPLOYED
     # ENVIRONMENTS. At the time of writing, Hoodoo ensures this by only using
     # debug logging in 'development' or 'test' environments.
     #
@@ -1678,15 +1681,15 @@ module Hoodoo; module Services
         # Rack provides no formal way to find out our host or port before a
         # request arrives, because in part it might change due to clustering.
         # For local development on an assumed single instance server, we can
-        # ask Ruby itself for all Rack::Server instances, expecting just one.
+        # ask Ruby itself for all Rackup::Server instances, expecting just one.
         # If there isn't just one, we rely on the Rack monkey patch or a
         # hard coded default.
 
         host = nil
         port = nil
 
-        if defined?( ::Rack ) && defined?( ::Rack::Server )
-          servers = ObjectSpace.each_object( ::Rack::Server )
+        if defined?( ::Rackup ) && defined?( ::Rackup::Server )
+          servers = ObjectSpace.each_object( ::Rackup::Server )
 
           if servers.count == 1
             server = servers.first
@@ -1706,7 +1709,7 @@ module Hoodoo; module Services
           port ||= '9292'
         end
 
-        # Announce the resource endpoints. We might not be able to annouce
+        # Announce the resource endpoints. We might not be able to announce
         # the remote availability of this endpoint if the host/port are not
         # determined; but that might just be because we are running under
         # "racksh" and we wouldn't want to announce remotely anyway.
@@ -1964,11 +1967,14 @@ module Hoodoo; module Services
       # data to read, it should return nil. If it doesn't, the payload is
       # too big. Reject it.
 
-      body = interaction.rack_request.body.read( MAXIMUM_PAYLOAD_SIZE )
+      body = nil
+      if !interaction.rack_request.body.nil?
+        body = interaction.rack_request.body.read( MAXIMUM_PAYLOAD_SIZE )
 
-      unless ( body.nil? || body.is_a?( ::String ) ) && interaction.rack_request.body.read( MAXIMUM_PAYLOAD_SIZE ).nil?
-        return response.add_error( 'platform.malformed',
-                                   'message' => 'Body data exceeds configured maximum size for platform' )
+        unless ( body.nil? || body.is_a?( ::String ) ) && interaction.rack_request.body.read( MAXIMUM_PAYLOAD_SIZE ).nil?
+          return response.add_error( 'platform.malformed',
+                                     'message' => 'Body data exceeds configured maximum size for platform' )
+        end
       end
 
       debug_log( interaction, 'Raw body data read successfully', body )
@@ -2156,7 +2162,7 @@ module Hoodoo; module Services
     # locale value.
     #
     # We support neither a list of preferences nor "qvalues", so if there is
-    # a list, we only take the first item; if there is a qvalue, we strip it
+    # a list, we only take the first item; if there is a value, we strip it
     # leaving just the language part, e.g. "en-gb".
     #
     # +interaction+:: Hoodoo::Services::Middleware::Interaction instance
@@ -2422,7 +2428,7 @@ module Hoodoo; module Services
       interaction.context.response.add_header( 'Content-Type', "#{ interaction.requested_content_type || 'application/json' }; charset=#{ interaction.requested_content_encoding || 'utf-8' }" )
     end
 
-    # Simplisitic CORS preflight handler.
+    # Simplistic CORS preflight handler.
     #
     # * http://www.w3.org/TR/cors/
     # * http://www.w3.org/TR/cors/#preflight-request
@@ -2758,7 +2764,7 @@ module Hoodoo; module Services
     #                 describing the current interaction.
     #
     # The interaction's request data will be updated with list parameter
-    # information if successul. The interaction's response data will be
+    # information if successful. The interaction's response data will be
     # updated with error information if anything is wrong.
     #
     def process_query_string( interaction )
@@ -2849,7 +2855,7 @@ module Hoodoo; module Services
     #                 #process_query_string.
     #
     # The interaction's request data will be updated with list parameter
-    # information if successul. The interaction's response data will be
+    # information if successful. The interaction's response data will be
     # updated with error information if anything is wrong.
     #
     def process_query_hash( interaction, query_hash )