+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 require 'webrick'
 require 'webrick/https'
+require 'rackup'
 
 # Set the correct environment for testing.
 
@@ -257,6 +258,7 @@ def spec_helper_start_svc_app_in_thread_for( app_class, use_ssl = false, app_opt
 
   Thread.start do
     app = Rack::Builder.new do
+      use Rack::RewindableInput::Middleware
       use Hoodoo::Services::Middleware unless app_options[:skip_hoodoo_middleware]
       run app_class.new
     end
@@ -286,7 +288,7 @@ def spec_helper_start_svc_app_in_thread_for( app_class, use_ssl = false, app_opt
     # this is the application which will also run a local DRb server.
 
     begin
-      Rack::Server.start( options )
+      Rackup::Server.start( options )
     rescue => e
       puts "TEST SERVER FAILURE: #{e.inspect}"
       puts e.backtrace